+++ Xamarin.Animation/Easing.cs
@@ -33,22 +33,22 @@ namespace Xamarin.Motion
 	
 	public static class Easing
 	{
-		public static readonly Func<float, float> Linear = x => x;
+		public static readonly Func<double, double> Linear = x => x;
 		
-		public static readonly Func<float, float> SinOut = x => (float)Math.Sin (x * Math.PI * 0.5f);
-		public static readonly Func<float, float> SinIn = x => 1.0f - (float)Math.Cos (x * Math.PI * 0.5f);
-		public static readonly Func<float, float> SinInOut = x => -(float)Math.Cos (Math.PI * x) / 2.0f + 0.5f;
+		public static readonly Func<double, double> SinOut = x => Math.Sin (x * Math.PI * 0.5f);
+		public static readonly Func<double, double> SinIn = x => 1.0f - Math.Cos (x * Math.PI * 0.5f);
+		public static readonly Func<double, double> SinInOut = x => -Math.Cos (Math.PI * x) / 2.0f + 0.5f;
 		
-		public static readonly Func<float, float> CubicIn = x => x * x * x;
-		public static readonly Func<float, float> CubicOut = x => (float)Math.Pow (x - 1.0f, 3.0f) + 1.0f;
-		public static readonly Func<float, float> CubicInOut = x => x < 0.5f ? (float)Math.Pow (x * 2.0f, 3.0f) / 2.0f :
-			(float)(Math.Pow ((x-1)*2.0f, 3.0f) + 2.0f) / 2.0f;
+		public static readonly Func<double, double> CubicIn = x => x * x * x;
+		public static readonly Func<double, double> CubicOut = x => Math.Pow (x - 1.0f, 3.0f) + 1.0f;
+		public static readonly Func<double, double> CubicInOut = x => x < 0.5f ? Math.Pow (x * 2.0f, 3.0f) / 2.0f :
+			(double)(Math.Pow ((x-1)*2.0f, 3.0f) + 2.0f) / 2.0f;
 		
-		public static readonly Func<float, float> BounceOut;
-		public static readonly Func<float, float> BounceIn;
+		public static readonly Func<double, double> BounceOut;
+		public static readonly Func<double, double> BounceIn;
 		
-		public static readonly Func<float, float> SpringIn = x => x * x * ((1.70158f + 1) * x - 1.70158f);
-		public static readonly Func<float, float> SpringOut = x => (x - 1) * (x - 1) * ((1.70158f + 1) * (x - 1) + 1.70158f) + 1;
+		public static readonly Func<double, double> SpringIn = x => x * x * ((1.70158f + 1) * x - 1.70158f);
+		public static readonly Func<double, double> SpringOut = x => (x - 1) * (x - 1) * ((1.70158f + 1) * (x - 1) + 1.70158f) + 1;
 		
 		static Easing ()
 		{
